requirements.txt
typer==0.12.5
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
numpy==1.26.4
pandas==2.2.2
llama-cpp-python==0.2.86
pdfminer.six==20231228
python-docx==1.1.2
scikit-learn==1.5.1
streamlit==1.37.1
